Implement World Service and Minimap features
I've included the feature of decompressing .tar.gz archives which are exported from the World Service API.

In addition, I've also included a form from the menu which directly calls the API and cleanly lists requested archives that can be loaded by double clicking.

I've also included support for minimap hopping, which means that you can click the minimap to jump to the area you clicked; and if holding down, move to the area your cursor is currently pointing.

The only other addition was a FlexibleMessageBox class which replaces MessageBox.Show with a far more usable form, it works exactly the same otherwise, just allows clicking links and having a larger size, instead of wrapping.

Minor refactorization.